Update pop_events.txt
There is one change on line 172 number 6 for a 5 if the pop is exactly 6 the event fail to modify the pop.
I remove almost all the "fire_only_once = yes", when running them the events on console all the trigger check pass but the code is not executed to fix the issue I change them with country_flags as per the information I read works the same as the "fire_only_once" but no idea why the "fire_only_once" is not letting the "immediate" run. I just wanted to share my solution for this so not all the changes that you have implementent are on my code  so please make a little check before merging. If you like to test the code use the base game and reduce "mean_time_to_happen = { months = 200 }" to less time. Hope it helps in your project and good luck with the rest of it.